% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_column_existence.R
\name{check_column_existence}
\alias{check_column_existence}
\title{Check if columns exist in a data frame}
\usage{
check_column_existence(df, col_names)
}
\arguments{
\item{df}{A data frame}

\item{col_names}{A vector of column names}
}
\value{
Boolean indicating if all specified column names exist in the data frame.
Otherwise, a message ascertaining non-existence of columns is displayed
}
\description{
Check if columns exist in a data frame
}
