% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_condition.R
\name{check_condition}
\alias{check_condition}
\title{Return values after applying a condition}
\usage{
check_condition(df, x, condition)
}
\arguments{
\item{df}{A data frame}

\item{x}{Character name of column}

\item{condition}{An anonymous function expressing a condition}
}
\value{
A vector of boolean values
}
\description{
Return values after applying a condition
}
\examples{
check_condition(cesd, "b1_row_missing", function(x) x == 1)
}
