% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_dir_existence.R
\name{check_dir_existence}
\alias{check_dir_existence}
\title{Check for existence of a directory.}
\usage{
check_dir_existence(...)
}
\arguments{
\item{...}{Strings of the names within a single path}
}
\value{
Boolean that indicates if directory exists, A message displays the path to the directory.
If path does not exist, an error message is displayed.
}
\description{
Highly recommend to use R Studio projects so paths are relative to the current working directory.
Wrapper to dir.exists
}
\examples{
check_dir_existence("data", "raw")
check_dir_existence("data", "processed")
check_dir_existence("docs")
}
