% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_summary_score.R
\name{compute_summary_score}
\alias{compute_summary_score}
\title{Compute sum and average scores}
\usage{
compute_summary_score(df, pattern, score_name, score_type = "average")
}
\arguments{
\item{df}{A data frame}

\item{pattern}{A regex string specifying which variables to compute summary scores on.}

\item{score_name}{The name of the score variable. Specify as string.}

\item{score_type}{The name of the score type. Currently supports: {average, sum}}
}
\value{
A data frame with new summary score column. The new column name is score_name
}
\description{
Compute sum and average scores
}
\examples{
compute_summary_score(stress_questionnaire, score_name = "physical_stress_average", score_type = "average")
}
