% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_factor_to_numeric.R
\name{convert_factor_to_numeric}
\alias{convert_factor_to_numeric}
\title{Convert a factor type to a numeric type
levels() gets the factor levels as characters
as.numeric() converts the factor levels to numeric
subsetting the factor column by as.numeric(levels()) gives the numeric form of the factor column}
\usage{
convert_factor_to_numeric(x)
}
\arguments{
\item{x}{A vector of type factor}
}
\value{
A vector of type numeric
}
\description{
Convert a factor type to a numeric type
levels() gets the factor levels as characters
as.numeric() converts the factor levels to numeric
subsetting the factor column by as.numeric(levels()) gives the numeric form of the factor column
}
\examples{
scores <- factor(x = c("1", "2"))
factor_to_numeric(scores)
}
