% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_row_non_missing.R
\name{count_row_non_missing}
\alias{count_row_non_missing}
\title{count number of non-missing entries}
\usage{
count_row_non_missing(df, pattern, col_name = "count_non_missing")
}
\arguments{
\item{df}{A data frame}

\item{pattern}{A regex string to subset columns.}

\item{col_name}{Column name of the number of non-missing values for each row}
}
\value{
A data frame with new column containing the count of non-missing values for each column. By default, the new column
is called count_row_missing
}
\description{
count number of non-missing entries
}
\examples{
count_row_missing_values(mtcars)
}
