% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_total_missing_values.R
\name{count_total_missing_values}
\alias{count_total_missing_values}
\title{Count total number of missing values in a data frame.}
\usage{
count_total_missing_values(df)
}
\arguments{
\item{df}{A data frame.}
}
\value{
Count of missing values in the data frame
}
\description{
Count total number of missing values in a data frame.
}
\examples{
count_total_missing_values(mtcars)
}
