% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_dir.R
\name{create_dir}
\alias{create_dir}
\title{Create a directory.}
\usage{
create_dir(...)
}
\arguments{
\item{...}{Strings of the names within a single path}
}
\value{
Create a new directory at the specified path
}
\description{
Highly recommend to use R Studio projects so paths are relative to the current working directory.
Wrapper to dir.create.
}
\examples{
create_dir("data", "raw")
create_dir("data", "processed")
}
