% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_similar_vars.R
\name{generate_similar_vars}
\alias{generate_similar_vars}
\title{Create a vector of similar variable names}
\usage{
generate_similar_vars(base_exp, affix, prefix = NULL, suffix = NULL)
}
\arguments{
\item{base_exp}{String. Common phrase among variables}

\item{affix}{Vector. Unique phrase to attach to base_exp}

\item{prefix}{Boolean. Append attachment to the head of base_exp}

\item{suffix}{Boolean. Append attachment to the tail of base_exp}
}
\value{
A vector of similar variable names. All variables names contain the affix.
}
\description{
Create a vector of similar variable names
}
\examples{
generate_similar_vars(base_exp = "m1_blood", affix = 1:10, suffix = TRUE)
generate_similar_vars(base_exp = "m1_blood", affix = c("weight", "height"), prefix = TRUE)
generate_similar_vars(base_exp = c("t0_month", "t1_month"), affix = 1:12, suffix = TRUE)
}
