% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_common_var.R
\name{get_common_var}
\alias{get_common_var}
\title{Get common variables from a list of data frames}
\usage{
get_common_var(dfs)
}
\arguments{
\item{dfs}{A list of data frames}
}
\value{
A string or vector of strings containing the common variables.
}
\description{
Get common variables from a list of data frames
}
\examples{
child <- data.frame(child_id = 4, age = 12)
parent <- data.frame(parent_id = 3, age = 34)
get_common_var(list(child, parent))
}
