% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_enviornment_vars.R
\name{get_environment_vars}
\alias{get_environment_vars}
\title{Get environment variables which match a regex pattern}
\usage{
get_environment_vars(pattern)
}
\arguments{
\item{pattern}{A string representing the pattern to search for.}
}
\value{
A named vector of environment variables which match the pattern
}
\description{
Get environment variables which match a regex pattern
}
\examples{
get_environment_vars(pattern = "^REDCAP")
get_environment_vars(pattern = "API_TOKEN")
}
