% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute_value.R
\name{impute_value}
\alias{impute_value}
\title{Impute a specified value to a replacement value if values of the rest of the row is NA}
\usage{
impute_value(data, impute_col, var_prefix, indices, replacement = NA)
}
\arguments{
\item{data}{a data frame}

\item{impute_col}{column in data to perform imputation on. supply as character string}

\item{var_prefix}{character vector of shared variable name}

\item{vars_indices}{integer indices attached to var_prefix. Used to subset columns when
checking for NA values.}
}
\value{
a data frame. impute_col will have imputed values
}
\description{
Impute a specified value to a replacement value if values of the rest of the row is NA
}
\examples{
impute_value(m1_cesd, m1_cesd_sum_score, var_prefix = "m1_cesd", indices = cesd_indices, replacement = NA)
}
