% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_nan.R
\name{is_nan}
\alias{is_nan}
\title{Checks for any NaN values in each column of a data frame}
\usage{
is_nan(x)
}
\arguments{
\item{x}{A data frame}
}
\value{
A matrix of logical indicating if a entry is NaN
}
\description{
Meant for use with logical subsetting to convert all NaN values in a data
frame to NA
}
\examples{
df <- data.frame(a = c(NaN, 1, 2), b = c(1, 2, NaN))
is_nan(df)
}
