% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join_multiple_data_frame.R
\name{join_multiple_data_frame}
\alias{join_multiple_data_frame}
\title{Join a list of data frame}
\usage{
join_multiple_data_frame(dfs, key, join_type)
}
\arguments{
\item{dfs}{A list of data frames.}

\item{key}{A character vector to specify which column(s) to use as join key(s).}

\item{join_type}{A string that specifies the join type. The options are:
inner, left, right, outer}
}
\value{
A single data frame after joining each individual data frame in dfs.
}
\description{
A wrapper to Reduce and merge in base R. Generalize to handle a list of data frames.
Currently, there is an assumption that the name of the key column is the same for all data frames in the list
}
