% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/order_columns.R
\name{order_columns}
\alias{order_columns}
\title{Sort a data frame by column name}
\usage{
order_columns(df, lead_vars)
}
\arguments{
\item{df}{A data frame}

\item{lead_vars}{A vector of variable names. Variables are put at the beginning
and not included in the sorting operation.}
}
\value{
A data frame, sorted by column names and prefixed by lead_vars.
}
\description{
Sort a data frame by column name
}
\examples{
order_columns(hospital, c("patient_id", "hospital_id"))
}
