% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_redcap_metadata.R
\name{read_redcap_metadata}
\alias{read_redcap_metadata}
\title{Read in metadata from Redcap project via Redcap API}
\usage{
read_redcap_metadata(
  uri,
  token,
  fields = NULL,
  forms = NULL,
  list_tokens = NULL
)
}
\arguments{
\item{uri}{A link to redcap server}

\item{token}{A string granting access to a REDCAP project via the REDCAP API}

\item{fields}{A string or a vector of strings. Indicates desired fields to query from REDCAP project.}

\item{forms}{A string or a vector of strings. Indicates desired forms to query from REDCAP project.}

\item{list_tokens}{A list of tokens granting accessing to multiple REDCAP projects via the REDCAP API}
}
\value{
A data frame containing metadata about a Redcap project. Four columns are returned:
field_name, form_name, field_label, and select_choices_or_calculations.
}
\description{
Read in metadata from Redcap project via Redcap API
}
\examples{
pregnancy_metadata <- map(PregnancyTokens, ~ReadMetadata(token = .x, list_tokens = PregnancyTokens))
}
