% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_redcap_project.R
\name{read_redcap_project}
\alias{read_redcap_project}
\title{Read in data from Redcap project via Redcap API}
\usage{
read_redcap_project(
  uri,
  token,
  fields = NULL,
  forms = NULL,
  raw_or_label = "raw",
  raw_or_label_headers = "raw",
  list_tokens = NULL
)
}
\arguments{
\item{uri}{A link to redcap server}

\item{token}{A string granting access to a REDCAP project via the REDCAP API}

\item{fields}{A string or a vector of strings. Indicates desired fields to query from REDCAP project.}

\item{forms}{A string or a vector of strings. Indicates desired forms to query from REDCAP project.}

\item{raw_or_label}{A string representing whether to include raw or labelled variable values. Two options: "raw" or "label"}

\item{raw_or_label_headers}{A string representing whether to include raw or labelled variable names. Two options: "raw" or "label".}

\item{list_tokens}{A list of tokens granting accessing to multiple REDCAP projects via the REDCAP API}
}
\value{
A data frame containing Redcap project data
}
\description{
Read in data from Redcap project via Redcap API
}
\examples{
ReadProjects(token = MRABToken, raw_or_label = "label", raw_or_label_tokens = "label")
pregnancy_data <- map(PregnancyTokens, ~ReadProjects(token = .x, list_tokens = PregnancyTokens))
infant_data <- map(InfantTokens, ~ReadProjects(token = .x, list_tokens = InfantTokens))
}
