% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename_column.R
\name{rename_column}
\alias{rename_column}
\title{Rename column}
\usage{
rename_column(df, old_name, new_name)
}
\arguments{
\item{df}{A data frame}

\item{old_name}{Original column name. Single element character vector.}

\item{new_name}{New column name. Single element character vector.}
}
\value{
A data frame with renamed column.
}
\description{
Rename column
}
\examples{
rename_column(iris, "Species", "species")
}
