% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace_on_condition.R
\name{replace_on_condition}
\alias{replace_on_condition}
\title{Mutate column given a condition}
\usage{
replace_on_condition(
  df,
  condition,
  pattern_condition_col,
  replacement_value,
  pattern_replacement_cols
)
}
\arguments{
\item{df}{A data frame}

\item{condition}{An anonymous Function that evaluates to a boolean.}

\item{pattern_condition_col}{A regex pattern that specifies the column to test the condition. As of now, limit
to only one condition column}

\item{replacement_value}{Replacement value if condition is true.}

\item{pattern_replacement_cols}{A regex pattern that specifies the columns to perform replacement}
}
\value{
A data frame
}
\description{
Mutate column given a condition
}
