% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset_row.R
\name{subset_row}
\alias{subset_row}
\title{Subset a data frame row-wise by matching row values to a regex pattern}
\usage{
subset_row(df, pattern, x)
}
\arguments{
\item{df}{A data frame}

\item{pattern}{A regex pattern.}

\item{x}{Column name to search through. Only specify 1 column name}
}
\value{
A data frame with row values which match the pattern. All columns in the data frame are returned.
}
\description{
Subset a data frame row-wise by matching row values to a regex pattern
}
\examples{
subset_row(who_data, pattern = "^A", "country_name")
}
