% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_summary_score.R
\name{compute_summary_score}
\alias{compute_summary_score}
\title{Compute sum and average scores}
\usage{
compute_summary_score(
  df,
  var_affix,
  score_name,
  score_type = "average",
  suffix = "_score"
)
}
\arguments{
\item{df}{A data frame}

\item{var_affix}{A string forming the variable name pattern. Specifies which columns to use to compute a score}

\item{score_name}{The name of the score variable}

\item{score_type}{The name of the score type. Currently supports: {average, sum}}

\item{suffix}{Informative suffix to append to variable name.}
}
\value{
A data frame with new column of subscale scores. The new column name is concatanation of score_name and "_score"
}
\description{
Compute sum and average scores
}
\examples{
compute_summary_score(stress_questionnaire, var_affix = "physical_stress", score_name = "physical_stress_sum")
}
