% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute_value.R
\name{impute_value}
\alias{impute_value}
\title{Impute values to a replacement value if a condition is true}
\usage{
impute_value(df, impute_col, condition, replacement)
}
\arguments{
\item{df}{A data frame}

\item{impute_col}{A character vector specifying the column to perform imputation.}

\item{condition}{A logical condition.}

\item{replacement}{The replacement value if condition is true.}
}
\value{
A data frame. The selected columns that is subsetted by var_affix contain changed values
given the condition and replacement values.
}
\description{
Impute values to a replacement value if a condition is true
}
\examples{
impute_value(df = iris, impute_col = "Species", condition = iris$species == "setosa", replacement = "Setosa")
}
