% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_redcap_project.R
\name{read_redcap_project}
\alias{read_redcap_project}
\title{Read in data from Redcap project via Redcap API}
\usage{
read_redcap_project(
  uri,
  token_name,
  fields = NULL,
  forms = NULL,
  raw_or_label = "raw",
  raw_or_label_headers = "raw"
)
}
\arguments{
\item{uri}{Name of the environment variable that stores the link to a Redcap server.}

\item{token_name}{Name of the environment variable that stores the API key which grants access to a REDCAP project.}

\item{fields}{A string or a vector of strings. Indicates desired fields to query from REDCAP project.}

\item{forms}{A string or a vector of strings. Indicates desired forms to query from REDCAP project.}

\item{raw_or_label}{A string representing whether to include raw or labelled variable values. Two options: "raw" or "label"}

\item{raw_or_label_headers}{A string representing whether to include raw or labelled variable names. Two options: "raw" or "label".}
}
\value{
A data frame containing Redcap project data
}
\description{
Read in data from Redcap project via Redcap API
}
\examples{
read_redcap_project(uri = "path/to/Redcap/API/Server", token_name = "NAME_OF_token_name_IN_RENVIRON", raw_or_label = "label", raw_or_label_token_names = "label")
mapply(read_redcap_metadata, token = get_environment_vars("^REDCAP"), uri = "path/to/Redcap/API/Server")

}
